fix(stripe-webhooks): Fix deadlock errors (#5878)
Stripe webhook jobs — `PaymentProviders::Stripe::HandleEventJob`
(`customer.updated` / `setup_intent.succeeded`) and
`PaymentProviders::Stripe::Customers::FetchDefaultPaymentMethodJob` (at
customer creation) — were dead-lettering with
`ActiveRecord::Deadlocked`. They all set the customer's default payment
method through the shared `PaymentMethods::SetAsDefaultService`, which
demotes the customer's other payment methods and promotes the target
within one transaction. When two events for the same customer are
processed concurrently, the transactions lock the same `payment_methods`
rows in opposite order and deadlock; with `retry: 0` and no matching
`retry_on`, the losing job dies and its update never lands.

- Serialize the default flip per customer: `SetAsDefaultService` now
runs inside a `Customers::LockService` advisory lock (new
`:payment_method` scope), so concurrent same-customer flips can no
longer interleave.
- `UpdatePaymentMethodService` now writes the `StripeCustomer` settings
and the default payment-method record in a single transaction, so a
failure can't leave them out of sync.
- `HandleEventJob` and `FetchDefaultPaymentMethodJob` now `retry_on
Customers::FailedToAcquireLock, ActiveRecord::Deadlocked`, so lock
contention / deadlock retries instead of dead-lettering.
