+++ app/services/billable_metrics/aggregations/count_service.rb
@@ -36,7 +36,7 @@ module BillableMetrics
         return empty_results if aggregations.blank?
 
         result.aggregations = aggregations.map do |aggregation|
-          group_result = BaseService::Result.new
+          group_result = BillableMetrics::Aggregations::BaseService::Result.new
           group_result.grouped_by = aggregation.groups
           group_result.aggregation = aggregation.value
           group_result.count = aggregation.events_count
