+++ app/services/billable_metrics/aggregations/max_service.rb
@@ -35,7 +35,7 @@ module BillableMetrics
         return empty_results if aggregations.blank?
 
         result.aggregations = aggregations.map do |aggregation|
-          group_result = BaseService::Result.new
+          group_result = BillableMetrics::Aggregations::BaseService::Result.new
           group_result.grouped_by = aggregation.groups
           group_result.aggregation = aggregation.value
           group_result.options = options
