+++ app/services/billable_metrics/aggregations/weighted_sum_service.rb
@@ -64,7 +64,7 @@ module BillableMetrics
         end
 
         result.aggregations = aggregations.map do |aggregation|
-          group_result = BaseService::Result.new
+          group_result = BillableMetrics::Aggregations::BaseService::Result.new
           group_result.grouped_by = aggregation.groups
 
           group_result.aggregation = aggregation.value
