+++ config/application.rb
@@ -3,7 +3,6 @@
 require_relative "boot"
 
 require "rails/all"
-require "ostruct"
 
 Bundler.require(*Rails.groups)
