+++ spec/services/base_service_spec.rb
@@ -12,20 +12,6 @@ RSpec.describe ::BaseService do
 
   it { is_expected.to use_middleware(Middlewares::LogTracerMiddleware) }
 
-  context "with current_user" do
-    it "assigns the current_user to the result" do
-      result = described_class.new.send :result
-
-      expect(result).to be_a(::BaseService::Result)
-    end
-
-    it "does not assign the current_user to the result if it isn't a User" do
-      result = described_class.new([]).send :result
-
-      expect(result.user).to be_nil
-    end
-  end
-
   context "with activity_loggable" do
     let(:service_class) do
       action = activity_loggable_action
